requirements.txt
streamlit>=1.31.0
anthropic>=0.18.0
python-docx>=1.1.0
